sauvegarde temp avec whisper et jeu en cours

+++ src/java/controlleur/recordMot.sh
@@ -1,15 +1,20 @@
 #!/bin/sh
 
-if [ -d src/WAV/ ];
+if [ -d src/ressources/WAV/ ]; 
 then
 cd src/ressources/WAV
 else
 mkdir src/ressources/WAV
-cd src/WAV
+cd src/ressources/WAV
 fi
 
+if [ "$pwd" | tail -c4 = "WAV" ];
+then
+cat Hello
 rm *
+fi
 
 NOW=$(date +"%Y%m%d_%H%M%S.wav")
 echo $NOW
 arecord -d 5 -r 16000 -f S16_LE "$NOW"
+
